nodejs-bundle: remove build-depend on make
An excerpt from lintians output:

N:    The package declares a depends on a build essential package without
N:    using a versioned depends. Packages do not have to build-depend on any
N:    package included in build-essential. It is the responsibility of anyone
N:    building packages to have all build-essential packages installed. The
N:    only reason for an explicit dependency on a package included in
N:    build-essential is if a particular version of that package is required,
N:    in which case the dependency should include the version.