Revert "Remove "Conflicts: xorg-driver-binary" and "Replaces: xorg-driver-binary""
This reverts commit d7b68e16b78eec2e7cfe4fa1dce63a17ff854589.

Doing this is trickier because nvidia-libopencl1-* and nvidia-opencl-icd-*
packages also conflict with each other (between 304- and 331-versions),
and they conflict "more badly" because in both cases there are same
filepaths in different package versions ---> resolving the conflict
would mean more nasty trickery.  We'll see later what we are going to do.